+++ core/frontend/src/internal/cross-package.ts
@@ -11,3 +11,5 @@ export { AnimationNodeId } from "../common/internal/render/AnimationNodeId";
 export { GltfDataType, type GltfMeshPrimitive } from "../common/gltf/GltfSchema";
 export { OnScreenTarget, Target } from "./render/webgl/Target";
 export { PerformanceMetrics } from "./render/webgl/PerformanceMetrics";
+export { GLTimerResult, RenderDiagnostics, RenderSystemDebugControl } from "./render/RenderSystemDebugControl";
+export { formatAnimationBranchId } from "./render/AnimationBranchState";
